Switched to new local and base executor (#362)
* Switched to new local and base executor, added wait to get_result, and updated changelog

* minor fix to changelog
