+++ covalent/executor/executor_plugins/local.py
@@ -27,9 +27,11 @@ This is a plugin executor module; it is loaded if found and properly structured.
 import io
 import os
 from contextlib import redirect_stderr, redirect_stdout
+from multiprocessing import Queue as MPQ
 from typing import Any, Dict, List
 
 # Relative imports are not allowed in executor plugins
+from covalent._results_manager.result import Result
 from covalent._shared_files import logger
 from covalent._shared_files.util_classes import DispatchInfo
 from covalent._workflow.transport import TransportableObject
@@ -53,16 +55,47 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
 class LocalExecutor(BaseExecutor):
     """
     Local executor class that directly invokes the input function.
+
+    Args:
+        log_stdout: The path to the file to be used for redirecting stdout.
+        log_stderr: The path to the file to be used for redirecting stderr.
+        cache_dir: The location used for cached files in the executor.
+        kwargs: Key-word arguments to be passed to the parent class (BaseExecutor)
     """
+
+    def __init__(
+        self,
+        log_stdout: str = "stdout.log",
+        log_stderr: str = "stderr.log",
+        cache_dir: str = os.path.join(
+            os.environ.get("XDG_CACHE_HOME") or os.path.join(os.environ["HOME"], ".cache"),
+            "covalent",
+        ),
+        **kwargs,
+    ):
+        self.log_stdout = log_stdout
+        self.log_stderr = log_stderr
+        self.cache_dir = cache_dir
+
+        base_kwargs = {"cache_dir": self.cache_dir}
+        for key in kwargs:
+            if key in [
+                "conda_env",
+                "current_env_on_conda_fail",
+            ]:
+                base_kwargs[key] = kwargs[key]
+
+        super().__init__(**base_kwargs)
 
     def execute(
         self,
         function: TransportableObject,
         args: List,
         kwargs: Dict,
+        info_queue: MPQ,
+        task_id: int,
         dispatch_id: str,
         results_dir: str,
-        node_id: int = -1,
     ) -> Any:
         """
         Executes the input function and returns the result.
@@ -72,10 +105,12 @@ class LocalExecutor(BaseExecutor):
                       is ultimately returned by this function.
             args: List of positional arguments to be used by the function.
             kwargs: Dictionary of keyword arguments to be used by the function.
+            info_queue: A multiprocessing Queue object used for shared variables across
+                processes. Information about, eg, status, can be stored here.
+            task_id: The ID of this task in the bigger workflow graph.
             dispatch_id: The unique identifier of the external lattice process which is
                          calling this function.
             results_dir: The location of the results directory.
-            node_id: The node ID of this task in the bigger workflow graph.
 
         Returns:
             output: The result of the executed function.
@@ -85,27 +120,34 @@ class LocalExecutor(BaseExecutor):
         fn = function.get_deserialized()
         fn_version = function.python_version
 
+        exception = None
+
+        info_dict = {"STATUS": Result.RUNNING}
+        info_queue.put_nowait(info_dict)
+
         with self.get_dispatch_context(dispatch_info), redirect_stdout(
             io.StringIO()
         ) as stdout, redirect_stderr(io.StringIO()) as stderr:
 
-            app_log.warning(self.conda_env)
-
+            result = None
             if self.conda_env != "":
-                result = None
 
-                result = self.execute_in_conda_env(
+                result, exception = self.execute_in_conda_env(
                     fn,
                     fn_version,
                     args,
                     kwargs,
                     self.conda_env,
                     self.cache_dir,
-                    node_id,
+                    info_queue,
+                    task_id,
                 )
 
             else:
-                result = fn(*args, **kwargs)
+                try:
+                    result = fn(*args, **kwargs)
+                except Exception as e:
+                    exception = e
 
         self.write_streams_to_file(
             (stdout.getvalue(), stderr.getvalue()),
@@ -114,4 +156,23 @@ class LocalExecutor(BaseExecutor):
             results_dir,
         )
 
-        return (result, stdout.getvalue(), stderr.getvalue())
+        info_dict = info_queue.get()
+        info_dict["STATUS"] = Result.FAILED if result is None else Result.COMPLETED
+        info_queue.put(info_dict)
+
+        return (result, stdout.getvalue(), stderr.getvalue(), exception)
+
+    def get_status(self, info_dict: dict) -> Result:
+        """
+        Get the current status of the task.
+
+        Args:
+            info_dict: a dictionary containing any neccessary parameters needed to query the
+                status. For this class (LocalExecutor), the only info is given by the
+                "STATUS" key in info_dict.
+
+        Returns:
+            A Result status object (or None, if "STATUS" is not in info_dict).
+        """
+
+        return info_dict.get("STATUS", Result.NEW_OBJ)
