+++ refactor/client_side/example_dispatch.py
@@ -18,32 +18,26 @@
 #
 # Relief from the License may be granted by purchasing a commercial license.
 
-
-import time
-
 import covalent as ct
-from refactor.executor.executor_plugins.local import LocalExecutor
-
-executor = LocalExecutor()
 
 
-@ct.electron(executor=executor)
+@ct.electron
 def task_1(x):
     return x**2
 
 
-@ct.electron(executor=executor)
+@ct.electron
 def subtask(a, b):
     return a**b
 
 
-@ct.electron(executor=executor)
+@ct.electron
 @ct.lattice
 def task_2(y, z):
     return subtask(y, z)
 
 
-@ct.lattice(executor=executor)
+@ct.lattice
 def workflow(a):
 
     task_2(a, 10)
@@ -55,6 +49,14 @@ dispatch_id = ct.dispatch(workflow)(3)
 
 print(dispatch_id)
 
+result = ct.get_result(dispatch_id=dispatch_id, wait=True)
+
+print(result)
+
+
+print("Now using dispatch_sync: ")
+
+print(ct.dispatch_sync(workflow)(2))
 # time.sleep(3)
 
 # No matter what dispatch id is sent, it returns from the last one only
